+++ src/parse/software_task.rs
@@ -9,11 +9,13 @@ use crate::{
 impl SoftwareTask {
     pub(crate) fn parse(args: SoftwareTaskArgs, item: ItemFn) -> parse::Result<Self> {
         let valid_signature =
-            util::check_fn_signature(&item) && util::type_is_unit(&item.sig.output);
+            util::check_fn_signature(&item, true) && util::type_is_unit(&item.sig.output);
 
         let span = item.sig.ident.span();
 
         let name = item.sig.ident.to_string();
+
+        let is_async = item.sig.asyncness.is_some();
 
         if valid_signature {
             if let Some((context, Ok(inputs))) = util::parse_inputs(item.sig.inputs, &name) {
@@ -27,6 +29,7 @@ impl SoftwareTask {
                     inputs,
                     stmts: item.block.stmts,
                     is_extern: false,
+                    is_async,
                 });
             }
         }
@@ -34,7 +37,7 @@ impl SoftwareTask {
         Err(parse::Error::new(
             span,
             &format!(
-                "this task handler must have type signature `fn({}::Context, ..)`",
+                "this task handler must have type signature `(async) fn({}::Context, ..)`",
                 name
             ),
         ))
@@ -47,11 +50,13 @@ impl SoftwareTask {
         item: ForeignItemFn,
     ) -> parse::Result<Self> {
         let valid_signature =
-            util::check_foreign_fn_signature(&item) && util::type_is_unit(&item.sig.output);
+            util::check_foreign_fn_signature(&item, true) && util::type_is_unit(&item.sig.output);
 
         let span = item.sig.ident.span();
 
         let name = item.sig.ident.to_string();
+
+        let is_async = item.sig.asyncness.is_some();
 
         if valid_signature {
             if let Some((context, Ok(inputs))) = util::parse_inputs(item.sig.inputs, &name) {
@@ -65,6 +70,7 @@ impl SoftwareTask {
                     inputs,
                     stmts: Vec::<Stmt>::new(),
                     is_extern: true,
+                    is_async,
                 });
             }
         }
@@ -72,7 +78,7 @@ impl SoftwareTask {
         Err(parse::Error::new(
             span,
             &format!(
-                "this task handler must have type signature `fn({}::Context, ..)`",
+                "this task handler must have type signature `(async) fn({}::Context, ..)`",
                 name
             ),
         ))